Rename `misc.ts` to `helpers.ts`
