+++ src/typescript/sdk/src/utils/arena/index.ts
@@ -1,1 +1,1 @@
-export * from "./misc";
+export * from "./helpers";
